refactor(engine): move CoordinatorSpawnContext to types.ts with explicit fields, no index signature
Removes the index signature that made CoordinatorSpawnContext structurally
identical to Readonly<Record<string,unknown>>. All coordinator-passable fields
are now explicit (assembledContextSummary, pitchPath, prUrl, findings, severity,
shapingComplete, auditComplete). Passing unknown keys is now a compile error.
Cast to generic map only at the coordinator-deps.ts boundary where typed context
widens to WorkflowTrigger.context.
